Preserve headers that were present before the legacy application is run (possible BC BREAK)
This helps in the case that headers (in particular, a
session-id cookie) are set before the legacy application is dispatched.

Until now, these headers were treated as being part of the
legacy application's response. In case that response is not fully
used later on (for example, only parts of the response body are relevant)
these headers got lost.

With this change, we're trying to separate the "greenfield" headers (outside
legacy) and the headers that are generated by legacy.

We're pretty much confident right now that this is actually a bugfix
and the right thing to do (TM), but of course it is a subtle change in
behaviour that might make a difference your your use case.

YMMV, good luck!